Refactor History compoenent && add x icon image
