+++ src/components/History/index.module.sass
@@ -30,7 +30,7 @@
   align-items: center
   flex-direction: row
   &:hover
-    background-color: grey
+    background-color: #b0ceea
   p
     padding: 16px 16px
     width: 80%
